Hotfix

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int x = 0;
+	size_t x = 0;
 
 	while (h != 0)
 	{
